refactor(data): fold neurons-sync into data-api.mjs, drop the extra Worker
The neurons/neuron_daily write path (POST /api/v1/internal/neurons-sync)
targeted the IDENTICAL Postgres instance workers/data-api.mjs already reads
from -- splitting it into its own dedicated Worker bought nothing (both need
the same postgres.js driver either way) while adding a redundant deploy
pipeline, wrangler config, service binding, and secret for zero
bundle-budget benefit. That split is only justified for
registry-sync-api.mjs, which targets a genuinely SEPARATE Postgres instance,
deliberately isolated so a bug in one can't take the other down.

- Deleted workers/neurons-sync-api.mjs + wrangler.neurons-sync.jsonc.
- Moved the write route (auth check + upsert transaction) into
  workers/data-api.mjs as handleNeuronsSync, checked before the read path's
  GET-only method gate. Its own writeJson() response helper avoids leaking
  the read routes' `cache-control: public, max-age=10` onto a write ack.
- workers/api.mjs's handleNeuronsSyncProxy now forwards through the existing
  DATA_API service binding instead of a new NEURONS_SYNC_API one; the
  binding is removed from wrangler.jsonc.
- wrangler.data.jsonc documents the new NEURONS_SYNC_SECRET this Worker
  needs (`wrangler secret put ... -c wrangler.data.jsonc`).
- Ported all neurons-sync-api.test.mjs coverage into tests/data-api.test.mjs,
  upgrading its postgres.js mock to also support the sql(rows, ...columns)
  bulk-insert helper (data-api.mjs's existing tests never needed it).
- Also fixes a real CI failure: refresh-metagraph.yml's new sync step used
  `continue-on-error: true`, which validate-workflows.mjs bans repo-wide as
  failure-masking. Replaced with a script-level `|| echo "::warning::..."`
  that keeps the same "never block the D1 path" property without hiding it
  from the workflow's own success/failure signal.
- ADR 0014 and the affected header comments updated to match.
