+++ .github/workflows/refresh-metagraph.yml
@@ -94,16 +94,18 @@ jobs:
           CLOUDFLARE_ACCOUNT_ID: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
 
       # #4771 verification window: ALSO write the same unsigned snapshot to the
-      # new Postgres write path (workers/neurons-sync-api.mjs), alongside (not
-      # replacing) the R2/D1 step above. METAGRAPH_NEURONS_SOURCE stays "d1"
-      # until a live parity pass confirms Postgres agrees with D1 -- this step
-      # only populates that comparison, it does not change what's served.
-      # continue-on-error: a failure here must never fail the D1 path this
-      # workflow's whole job (and #1345's daily rollup cron) depends on. Skips
-      # cleanly (exit 0) until NEURONS_SYNC_SECRET is provisioned, same
-      # not-yet-configured convention as notify-on-failure's webhook guard below.
+      # new Postgres write path (workers/data-api.mjs's handleNeuronsSync),
+      # alongside (not replacing) the R2/D1 step above. METAGRAPH_NEURONS_SOURCE
+      # stays "d1" until a live parity pass confirms Postgres agrees with D1 --
+      # this step only populates that comparison, it does not change what's
+      # served. A failure here must never fail the D1 path this workflow's
+      # whole job (and #1345's daily rollup cron) depends on -- NOT
+      # `continue-on-error:` (banned repo-wide by validate-workflows.mjs as
+      # failure-masking); instead the script itself always exits 0, logging a
+      # failure loudly (::warning::) rather than swallowing it silently, and
+      # skips cleanly until NEURONS_SYNC_SECRET is provisioned (same
+      # not-yet-configured convention as notify-on-failure's webhook guard below).
       - name: Sync neurons to the Postgres write path (#4771, verification window)
-        continue-on-error: true
         run: |
           if [ -z "$NEURONS_SYNC_SECRET" ]; then
             echo "NEURONS_SYNC_SECRET not set; skipping Postgres sync"
@@ -112,7 +114,8 @@ jobs:
           curl -fsS -X POST "https://api.metagraph.sh/api/v1/internal/neurons-sync" \
             -H "x-neurons-sync-token: ${NEURONS_SYNC_SECRET}" \
             -H "content-type: application/json" \
-            --data-binary @dist/metagraph-neurons.json
+            --data-binary @dist/metagraph-neurons.json \
+            || echo "::warning::neurons-sync Postgres write failed; D1 path (this job's real dependency) is unaffected"
         env:
           NEURONS_SYNC_SECRET: ${{ secrets.NEURONS_SYNC_SECRET }}
