+++ workers/api.mjs
@@ -1161,24 +1161,29 @@ async function handleRegistrySyncProxy(request, env) {
   });
 }
 
-// Proxies POST /api/v1/internal/neurons-sync to the dedicated neurons-sync
-// Worker (NEURONS_SYNC_API service binding, #4771), the write path into the
-// chain-indexer Postgres's neurons/neuron_daily tables. Mirrors
-// handleRegistrySyncProxy exactly: forwards the request as-is (including the
-// x-neurons-sync-token header), the shared-secret check happens once,
-// downstream in workers/neurons-sync-api.mjs.
+// Proxies POST /api/v1/internal/neurons-sync to the SAME DATA_API service
+// binding the chain_events proxy above uses (#4771) -- the write path into
+// the chain-indexer Postgres's neurons/neuron_daily tables lives inside
+// workers/data-api.mjs itself, not a separate Worker: it's the identical
+// Postgres instance/Hyperdrive origin data-api.mjs already reads from, so
+// splitting read and write into two Workers (the way registry-sync-api.mjs
+// is split, for a genuinely SEPARATE database) would only add a redundant
+// deploy pipeline for zero bundle-budget benefit. Mirrors
+// handleRegistrySyncProxy's shape otherwise: forwards the request as-is
+// (including the x-neurons-sync-token header), the shared-secret check
+// happens once, downstream in workers/data-api.mjs's handleNeuronsSync.
 async function handleNeuronsSyncProxy(request, env) {
   if (request.method !== "POST") {
     return errorResponse("method_not_allowed", "Only POST is supported.", 405);
   }
-  if (!env.NEURONS_SYNC_API) {
+  if (!env.DATA_API) {
     return errorResponse(
       "neurons_sync_unavailable",
       "The neurons-sync tier is not bound to this deployment.",
       503,
     );
   }
-  const upstream = await env.NEURONS_SYNC_API.fetch(request);
+  const upstream = await env.DATA_API.fetch(request);
   let body;
   try {
     body = await upstream.json();
@@ -1302,9 +1307,10 @@ export async function handleRequest(request, env = {}, ctx = {}) {
   }
   // The write path into the chain-indexer Postgres's neurons/neuron_daily
   // tables (#4771) -- refresh-metagraph.yml's sign-and-stage job calls this
-  // over HTTPS alongside its existing R2-stage-to-D1 step. Proxies to the
-  // dedicated neurons-sync Worker (wrangler.neurons-sync.jsonc), same shape
-  // as the registry-sync proxy above.
+  // over HTTPS alongside its existing R2-stage-to-D1 step. Proxies to
+  // workers/data-api.mjs's handleNeuronsSync via the SAME DATA_API service
+  // binding the chain_events proxy above uses (not a separate Worker -- see
+  // handleNeuronsSyncProxy's comment for why).
   if (url.pathname === "/api/v1/internal/neurons-sync") {
     return handleNeuronsSyncProxy(request, env);
   }
